+++ packages/devtools_extensions/example/app_that_uses_foo/test/flutter_test_2.dart
@@ -1,4 +1,4 @@
-// Copyright 2023 The Chromium Authors. All rights reserved.
+// Copyright 2024 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -6,13 +6,18 @@ import 'package:app_that_uses_foo/main.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:foo/foo.dart';
 
-// This test can be run to verify that the `package:foo` DevTools extension
-// loads properly when debugging a test target with DevTools.
+// This test can be run to verify that the DevTools extensions available for
+// package:app_that_uses_foo load properly when debugging a Flutter test target
+// with DevTools.
 //
 // To test this, run the following command and copy the VM service URI to
 // connect to DevTools:
 //
-// flutter run test/app_that_uses_foo_test.dart --start-paused -d flutter-tester
+// flutter test test/flutter_test_2.dart --start-paused
+//
+// To test this test as part of a suite, use this command instead:
+//
+// flutter test test/ --start-paused
 
 void main() {
   testWidgets('Builds $MyAppThatUsesFoo', (tester) async {