+++ packages/devtools_app/lib/src/screens/memory/panes/tracing/tracing_data.dart
@@ -3,84 +3,158 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app_shared/utils.dart';
+import 'package:devtools_shared/devtools_shared.dart';
 import 'package:flutter/foundation.dart';
 import 'package:vm_service/vm_service.dart';
 
 import '../../../../shared/globals.dart';
 import '../../../../shared/memory/class_name.dart';
+import '../../../../shared/primitives/encoding.dart';
+import '../../../../shared/primitives/simple_items.dart';
 import '../../../../shared/primitives/utils.dart';
 import '../../../../shared/table/table_data.dart';
 import '../../../profiler/cpu_profile_model.dart';
 import '../../../profiler/cpu_profile_transformer.dart';
 
+@visibleForTesting
+enum TracedClassJson {
+  clazz,
+  instances,
+  allocations,
+}
+
 /// A representation of a class and it's allocation tracing state.
-class TracedClass with PinnableListEntry {
+class TracedClass with PinnableListEntry, Serializable {
   TracedClass({
-    required this.cls,
+    required this.clazz,
   })  : traceAllocations = false,
         instances = 0,
-        name = HeapClassName.fromClassRef(cls);
+        name = HeapClassName.fromClassRef(clazz);
 
   TracedClass._({
-    required this.cls,
+    required this.clazz,
     required this.instances,
     required this.traceAllocations,
-  }) : name = HeapClassName.fromClassRef(cls);
+  }) : name = HeapClassName.fromClassRef(clazz);
+
+  factory TracedClass.fromJson(Map<String, dynamic> json) {
+    return TracedClass._(
+      instances: json[TracedClassJson.instances.name] as int,
+      clazz: ClassRefEncodeDecode.instance
+          .decode(json[TracedClassJson.clazz.name]),
+      traceAllocations: json[TracedClassJson.allocations.name] as bool,
+    );
+  }
+
+  @override
+  Map<String, dynamic> toJson() {
+    return {
+      TracedClassJson.allocations.name: traceAllocations,
+      TracedClassJson.clazz.name: clazz,
+      TracedClassJson.instances.name: instances,
+    };
+  }
 
   TracedClass copyWith({
-    ClassRef? cls,
+    ClassRef? clazz,
     int? instances,
     bool? traceAllocations,
   }) {
     return TracedClass._(
-      cls: cls ?? this.cls,
+      clazz: clazz ?? this.clazz,
       instances: instances ?? this.instances,
       traceAllocations: traceAllocations ?? this.traceAllocations,
     );
   }
 
   final HeapClassName name;
-  final ClassRef cls;
+  final ClassRef clazz;
   final int instances;
   final bool traceAllocations;
 
   @override
   bool operator ==(Object other) {
     if (other is! TracedClass) return false;
-    return cls == other.cls &&
+    return clazz == other.clazz &&
         instances == other.instances &&
         traceAllocations == other.traceAllocations;
   }
 
   @override
-  int get hashCode => Object.hash(cls, instances, traceAllocations);
+  int get hashCode => Object.hash(clazz, instances, traceAllocations);
 
   @override
   bool get pinToTop => traceAllocations;
 
   @override
   String toString() =>
-      '${cls.name} instances: $instances trace: $traceAllocations';
+      '${clazz.name} instances: $instances trace: $traceAllocations';
+}
+
+@visibleForTesting
+enum TracingIsolateStateJson {
+  isolate,
+  classes,
+  profiles;
 }
 
 /// Contains allocation tracing state for a single isolate.
 ///
 /// `AllocationProfileTracingController` is effectively only used to provide
 /// consumers the allocation tracing state for the currently selected isolate.
-class TracingIsolateState {
-  TracingIsolateState({required this.isolate});
+class TracingIsolateState with Serializable {
+  TracingIsolateState({
+    required this.mode,
+    required this.isolate,
+    Map<String, CpuProfileData>? profiles,
+    List<TracedClass>? classes,
+    String? selectedClass,
+  }) {
+    this.classes = classes ?? [];
+    classesById = {for (var e in this.classes) e.clazz.id!: e};
+    this.profiles = profiles ?? {};
+  }
 
-  TracingIsolateState.empty() : isolate = IsolateRef();
+  TracingIsolateState.empty()
+      : this(isolate: IsolateRef(), mode: ControllerCreationMode.connected);
+
+  factory TracingIsolateState.fromJson(Map<String, dynamic> json) {
+    return TracingIsolateState(
+      mode: ControllerCreationMode.offlineData,
+      isolate: IsolateRefEncodeDecode.instance
+          .decode(json[TracingIsolateStateJson.isolate.name]),
+      profiles: (json[TracingIsolateStateJson.profiles.name] as Map).map(
+        (key, value) => MapEntry(
+          key,
+          deserialize<CpuProfileData>(value, CpuProfileData.fromJson),
+        ),
+      ),
+      classes: (json[TracingIsolateStateJson.classes.name] as List)
+          .map((e) => deserialize<TracedClass>(e, TracedClass.fromJson))
+          .toList(),
+    );
+  }
+
+  @override
+  Map<String, dynamic> toJson() {
+    return {
+      TracingIsolateStateJson.isolate.name: isolate,
+      TracingIsolateStateJson.classes.name: classesById.values.toList(),
+      TracingIsolateStateJson.profiles.name: profiles,
+    };
+  }
+
+  final ControllerCreationMode mode;
 
   final IsolateRef isolate;
 
   // Keeps track of which classes have allocation tracing enabling.
-  final tracedClasses = <String, TracedClass>{};
-  final tracedClassesProfiles = <String, CpuProfileData>{};
-  final unfilteredClassList = <TracedClass>[];
+  late final Map<String, TracedClass> classesById;
+  late final Map<String, CpuProfileData> profiles;
+  late final List<TracedClass> classes;
 
   /// The current class selection in the [AllocationTracingTable]
-  final selectedTracedClass = ValueNotifier<TracedClass?>(null);
+  final selectedClass = ValueNotifier<TracedClass?>(null);
 
   /// The list of classes for the currently selected isolate.
   ValueListenable<List<TracedClass>> get filteredClassList =>
@@ -91,8 +165,8 @@ class TracingIsolateState {
 
   /// The allocation profile data for the current class selection in the
   /// [AllocationTracingTable].
-  CpuProfileData? get selectedTracedClassAllocationData {
-    return tracedClassesProfiles[selectedTracedClass.value?.cls.id!];
+  CpuProfileData? get selectedClassProfile {
+    return profiles[selectedClass.value?.clazz.id!];
   }
 
   /// The last time, in microseconds, the table was cleared. This time is based
@@ -101,13 +175,20 @@ class TracingIsolateState {
   int _lastClearTimeMicros = 0;
 
   Future<void> initialize() async {
-    final classList = await serviceConnection.serviceManager.service!
-        .getClassList(isolate.id!);
-    for (final cls in classList.classes!) {
-      tracedClasses[cls.id!] = TracedClass(cls: cls);
+    if (mode == ControllerCreationMode.connected) {
+      final classList = await serviceConnection.serviceManager.service!
+          .getClassList(isolate.id!);
+      for (final clazz in classList.classes!) {
+        classesById[clazz.id!] = TracedClass(clazz: clazz);
+      }
+      classes.addAll(classesById.values);
+    } else {
+      for (final kv in profiles.entries) {
+        final profile = kv.value;
+        await _setProfile(classesById[kv.key]!, profile);
+      }
     }
-    _filteredClassList.replaceAll(tracedClasses.values);
-    unfilteredClassList.addAll(tracedClasses.values);
+    updateClassFilter('', force: true);
   }
 
   Future<void> refresh() async {
@@ -130,11 +211,11 @@ class TracingIsolateState {
     final updatedFilteredClassList =
         (newFilter.caseInsensitiveContains(currentFilter) && !force
                 ? _filteredClassList.value
-                : unfilteredClassList)
+                : classes)
             .where(
-              (e) => e.cls.name!.caseInsensitiveContains(newFilter),
+              (e) => e.clazz.name!.caseInsensitiveContains(newFilter),
             )
-            .map((e) => tracedClasses[e.cls.id!]!)
+            .map((e) => classesById[e.clazz.id!]!)
             .toList();
 
     _filteredClassList.replaceAll(updatedFilteredClassList);
@@ -147,34 +228,37 @@ class TracingIsolateState {
         (await serviceConnection.serviceManager.service!.getVMTimelineMicros())
             .timestamp!;
     // Reset the counts for traced classes.
-    final updatedTracedClasses = tracedClasses.map((key, value) {
+    final updatedTracedClasses = classesById.map((key, value) {
       return MapEntry(key, value.copyWith(instances: 0));
     });
 
-    tracedClasses
+    classesById
       ..clear()
       ..addAll(updatedTracedClasses);
 
     // Reset the unfiltered class list with the new `TracedClass` instances.
-    unfilteredClassList
+    classes
       ..clear()
-      ..addAll(tracedClasses.values);
+      ..addAll(classesById.values);
     updateClassFilter(currentFilter, force: true);
 
     // Since there's no longer any tracing data, clear the existing profiles.
-    tracedClassesProfiles.clear();
+    profiles.clear();
   }
 
-  /// Enables or disables tracing of allocations of [cls].
-  Future<void> setAllocationTracingForClass(ClassRef cls, bool enabled) async {
+  /// Enables or disables tracing of allocations of [clazz].
+  Future<void> setAllocationTracingForClass(
+    ClassRef clazz,
+    bool enabled,
+  ) async {
     final service = serviceConnection.serviceManager.service!;
     final isolate =
         serviceConnection.serviceManager.isolateManager.selectedIsolate.value!;
-    final tracedClass = tracedClasses[cls.id!]!;
+    final tracedClass = classesById[clazz.id!]!;
 
-    // Only update if the tracing state has changed for `cls`.
+    // Only update if the tracing state has changed for `clazz`.
     if (tracedClass.traceAllocations != enabled) {
-      await service.setTraceClassAllocation(isolate.id!, cls.id!, enabled);
+      await service.setTraceClassAllocation(isolate.id!, clazz.id!, enabled);
       final update = tracedClass.copyWith(
         traceAllocations: enabled,
       );
@@ -183,12 +267,12 @@ class TracingIsolateState {
   }
 
   void _updateClassState(TracedClass original, TracedClass updated) {
-    final cls = original.cls;
+    final clazz = original.clazz;
     // Update the currently selected class, if it's still being traced.
-    if (selectedTracedClass.value?.cls.id == cls.id) {
-      selectedTracedClass.value = updated;
+    if (selectedClass.value?.clazz.id == clazz.id) {
+      selectedClass.value = updated;
     }
-    tracedClasses[cls.id!] = updated;
+    classesById[clazz.id!] = updated;
     _filteredClassList.replace(original, updated);
   }
 
@@ -203,7 +287,7 @@ class TracingIsolateState {
     final service = serviceConnection.serviceManager.service!;
     final isolateId = serviceConnection
         .serviceManager.isolateManager.selectedIsolate.value!.id!;
-    final cls = tracedClass.cls;
+    final clazz = tracedClass.clazz;
 
     // Note: we need to provide `timeExtentMicros` to `getAllocationTraces`,
     // otherwise the VM will respond with all samples, not just the samples
@@ -212,7 +296,7 @@ class TracingIsolateState {
     // Request the allocation profile for the traced class.
     final trace = await service.getAllocationTraces(
       isolateId,
-      classId: cls.id!,
+      classId: clazz.id!,
       timeOriginMicros: _lastClearTimeMicros,
       timeExtentMicros: maxJsInt,
     );
@@ -222,6 +306,15 @@ class TracingIsolateState {
       cpuSamples: trace,
     );
 
+    await _setProfile(tracedClass, profileData);
+
+    return profileData;
+  }
+
+  Future<void> _setProfile(
+    TracedClass tracedClass,
+    CpuProfileData profileData,
+  ) async {
     // Process the allocation profile into a tree. We can reuse the transformer
     // from the CPU Profiler tooling since it also makes use of a `CpuSamples`
     // response.
@@ -232,10 +325,10 @@ class TracingIsolateState {
     final updated = tracedClass.copyWith(
       instances: profileData.cpuSamples.length,
     );
-    tracedClasses[cls.id!] = updated;
-    tracedClassesProfiles[cls.id!] = profileData;
+    final clazz = tracedClass.clazz;
+    classesById[clazz.id!] = updated;
+    profiles[clazz.id!] = profileData;
 
     _updateClassState(tracedClass, updated);
-    return profileData;
   }
 }