+++ packages/devtools_app/lib/src/shared/feature_flags.dart
@@ -51,7 +51,7 @@ abstract class FeatureFlags {
   /// Flag to enable widget rebuild stats ui.
   ///
   /// https://github.com/flutter/devtools/issues/4564.
-  static bool widgetRebuildStats = enableExperiments;
+  static bool widgetRebuildStats = true;
 
   /// Flag to enable the deep link validation tooling in DevTools, both for the
   /// DevTools screen and the standalone tool for IDE embedding.