+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.0-dev.1
+* Add testimonies from extension authors to the `README.md`.
+
 ## 0.2.0-dev.0
 * Deprecate the `DevToolsExtension.requiresRunningApplication` field.
 * Update `extension_config_spec.md` to include an optional field `requiresConnection`.